requirements.txt
asgiref==3.4.1
certifi==2023.5.7
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.8
Django==3.1.12
django-cors-headers==3.10.1
django-rest-framework-mongoengine==3.4.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangotoolbox==1.8.0
djongo==1.3.6
drf-yasg==1.21.6
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.4
MarkupSafe==2.0.1
martor==1.6.14
mongoengine==0.24.2
openapi-codec==1.3.2
packaging==21.3
pymongo==3.12.1
pyparsing==3.0.7
pytz==2023.3
PyYAML==6.0
requests==2.27.1
simplejson==3.19.1
sqlparse==0.2.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.16
zipp==3.6.0
